refactor!: add support for tokenize_batch method
BREAKING CHANGE: remove support for previous custom tokenization and its methods
